+++ safe-road-backend/module-gamification/src/main/java/ru/itmo/saferoad/gamification/service/GameProfileService.java
@@ -20,5 +20,5 @@ public interface GameProfileService {
 
 	List<LeaderboardProjection> getTotalTop10();
 
-	long getRank(@NotNull Integer xp);
+	long getRank(@NotNull Long xp);
 }
